Implement DirectoryNode and serialization
The files and directories of a snapshot are represented as a tree.
The serialization of this tree is a bit tricky: I want to keep the
complete tree in memory but serialize individual directory nodes to
file. To do that, I reference other nodes with a hash in the serialized
version.