Master key
Implemented a master key encryption system.
When the user inits a new repo, a new, random master key is created.
This key is encrypted using the user's password. The encrypted key and
other metadata are stored in a key file. Multiple key files can be
added to the keys folder, allowing multiple passwords (with the same
master key).