Remove accessed_time from Node Metadata
The accessed time is changed every time we scan a file, so the node's
will be different every time, completely destroying the benefit of
tree deduplication. This field will be updated when the file is restored
anyway, so we can just ignore it.

If anything, maybe we can just store it for completion, but ignore it
for calculating the hash of the Metadata struct. Ignore for now.