Reimplement Commiter with channels
The committer looks ugly with all those mutexes. It will be a nightmare
to maintain in the future. For the 6th time, I have changed the implementation,
this time using channels in a way that works. This implementation seems to have
exactly the same performance as before, but with much better code.