Extract key retrieval from repository logic
Thy key should be retrieved before opening a repository. In order to open the
repository, you need to use a key you have acquired beforehand. One doesn't simply
open a repository and then find a key once inside, right?

Having the key before opening the repository will allow us to create the SecureStorage
before the repo, and use it as a dependency of other things.