Refactor encryption and compression
To implement packs, I need a way to save objects to the repo which are
sometimes encoded [compressed + encrypted] before saving them (trees and
snapshots), and sometimes saved raw (packs) because the contents (the
blobs) are encoded already within the pack. The ID of the repo object is
the hash of the data saved to file (don't care if it's encoded or not).

To allow for such granularity, the SecureStorage cannot be used to save
and load files, but rather to encode and decode. The Repository can save
objects by hashing them first, or to save them raw. Either way, the hash
calculated and returned corresponds to the actual data saved to file.

There are now two pairs of functions in the Repository to save objects:

- save_object()
- save_object_raw()
- load_object()
- load_object_raw()

This is to save the producers of these objects from having to encode
the data.

With all this, the SecureStorage no longer has a reference to the
StorageBackend, since it doesn't actually save or load anything.