Stabilized paths
The Archiver stores paths relative to the LCP of the target paths. The
root path is store for future reference.

The FSNodeStreamer must not emit the root node, but it must emit the
intermediate nodes without exploring into them.
The SerializedNodeStreamer is consistent with this, since the Snapshot
tree does not contain a dedicated Node for the LCP.

However, for this to work, the Archiver needs to know how many root
children to expect. This requires calculating the intermediate paths in
two separate places. This calculation doesn't add a significant
overhead, but I should anyway consider if this can be optimized.