Refactor ID
Previously, all content-addressable objects in the repository were identified
with a String. This has the problem that Strings don't have fixed sizes.

ID is now a slice of bytes, the size of a hash (currently 256-bit BLAKE3). ID
can be converted to/from hex strings, but have a fix size with can be configured
if we ever change the hashing algorithm.