Sort nodes lexicographically in ls
The Archiver does not sort the nodes in the tree so we need to sort them after we
read the tree. A minor inconvenience, but I think it is better to do it like this
since the SerializedNodeStreamer should sort the paths anyway.

We do not care about the order of the nodes in the serialized tree as long as they
are sorted when consumed, which we must always enforce for correctness.