Flush pending indexes periodically
During the snapshot process, we flush the index periodically in order to
commit packs frequently. If the snapshot process is interrupted, the packs
referenced by an index are not lost when resuming the snapshot. However,
flushing often means saving a lot of small indexes to file. The garbage
collector should merge all small indexes and consolidate them into bigger
index files.