Add metadata to pack files
Pack files now contain a section at the end of the file that contains
the metadata describing the blob content.

The last 4 bytes of the filecontain the length of the section. For each blob
in the pack, the metadata section contains:

  - ID (32 bytes)
  - Length (4 bytes)
  - Blob type (1 byte)

The structure of the metadata section is:

  | id | len | type | id | len | type | ... | META SIZE |

This metadata should allow to scan the contents of the pack directly, without
reading the index. Also, if a pack became unreferenced from any snapshot, its
contents can still be processed.