Migrate from AES-GCM to AES-GCM-SIV
Switched to AES-GCM-SIV to gain nonce misuse resistance. This prevents
catastrophic security failures in the event of accidental IV reuse,
which is a risk when generating random IVs.

SecureStorage generates a random IV for every encoded message. The
probability of a collission is astronomically low, but still greater
than 0. The performance penalty of the SIV version should be negligible,
while protecting about the very edge case of reusing an IV.