Add locks and handle interrupts
Added a signal handler to allow cleanup when SIGINT or SIGTERM are
received.

Added exclusive and non-exclusive Locks and a LockHandle to refresh
the locks and allow deferred unlocking.

Added a defer! macro to defer execution of statements at the end of a
scope. Used to unlock the LockHandle when the command is finished or
during cleanup.

Added an unlock command to remove expired locks from interrupted
commands.