Add username to key files
Authentication now requires a username, apart from the password.
Since we are moving towards a multi-user repository, we want to
distinguish keyfiles that use the same user password. To do that, we
introduce usernames. To open a repository, the user must now input their
username and password. The username is used to ignore all keyfiles
belonging to other users.

This is a breaking change, since the old keyfiles can no longer be
parsed by the client.