Create key command
Added KeyManager to handle keys without opening a repository.

The key command allows listing existing Keyfiles (keyfiles in the keys
directories, as other keys can exist out of the repository), creating
new Keyfiles, deleting Keyfiles and changing the password of a Keyfile
associated to a username.

Master key rotation is not implemented yet.