Fail if no source paths are provided as arguments for cmd_snapshots
The argument parser will not allow passing an empty vector for the source
paths, but the command can still accept an empty vector.
We need to double check that args.paths is not empty in the command and
fail otherwise.

test_cmd_amend created an empty snapshot using the run method of the
snapshot command, which should never be allowed.