Best effort metadata restore and warning logs
Failing to set the owner should not be an error or crowd the console
with warnings, since it is an operation that is mostly expected to
fail during restore. As a best effort tool, mapache tries to chown,
but there is no guarantee.

In general, failing to restore metadata is not an error, but a warning.
Highly likely and recurrent warnings, which don't affect the integrity
of the data, should not be logged.