Enforce read_concurrency in Archiver processor thread
Install a pool of 'read_concurrency' rayon threads and use a parallel
iterator. This enforces that only 'read_concurrency' files are processed
at once.
This is a regression introduced by 8402b51839e425da479739be458a47aee68ff049.
We keep the parallel iterator, but restore the thread pool. We install the
pool instead of using a scope.