Include all non-directory nodes in the snapshot file count
If the target paths contain symlinks or other non-file/non-directory
nodes, the scanner task does not count them but the processor does.
At the end of the Archive task, the number of processed items is
greater than the number of scanned items, which can be seen as an error
by the user.